Move new paywall component logic to PaywallComponent. Update plans and types usage
