+++ js/app/packages/app/component/paywall/PaywallTeamOwnerView.tsx
@@ -1,16 +1,16 @@
+import { PLANS } from '@app/component/paywall/plans';
 import { useTeamQuery } from '@queries/team';
 import type { Team } from '@service-auth/generated/schemas/team';
 import { stripeServiceClient } from '@service-stripe/client';
 import { Button } from '@ui';
 import { createMemo } from 'solid-js';
-import { NEW_PLANS } from './plans';
 import SubscriptionTier from './SubscriptionTier';
 
 interface PaywallTeamOwnerViewProps {
   team: Team;
 }
 
-const PREMIUM_PRICE = NEW_PLANS.find((p) => p.tier === 'premium')!.price;
+const PREMIUM_PRICE = PLANS.find((p) => p.tier === 'premium')!.price;
 
 const PaywallTeamOwnerView = (props: PaywallTeamOwnerViewProps) => {
   const teamQuery = useTeamQuery(() => props.team.id);
